\name{constant_weight_function}
\alias{constant_weight_function}
\title{Constant weight function}
\usage{
constant_weight_function(d, dt)
}
\arguments{
  \item{d}{trajectory distance}

  \item{time}{distance}
}
\description{
Constant weight function for local regression, corresponds
roughly to a square function with adaptive bandwidth.
}

