\name{irregular_derivatives}
\alias{irregular_derivatives}
\title{Compute timestep derivatives in the abscence of regular timesteps}
\usage{
irregular_derivatives(tt)
}
\arguments{
  \item{tt}{\code{time.table} to estimate derivative of
  points in}
}
\description{
Compute timestep derivatives in the abscence of regular
timesteps
}

