\name{lookup_neighbour_trajectories}
\alias{lookup_neighbour_trajectories}
\title{Find neighbouring trajectories}
\usage{
lookup_neighbour_trajectories(tt, points, k, timesteps = NULL, times = NULL,
  only.indices = TRUE, trajectory.distance.name = "distance",
  time.distance.name = "relative")
}
\description{
Find neighbouring trajectories
}

