\name{simple_weight_function}
\alias{simple_weight_function}
\title{A simple and stupid weight function}
\usage{
simple_weight_function(d, dt, h = 0.1)
}
\arguments{
  \item{d}{trajectory distance}

  \item{dt}{time distance}

  \item{h}{gaussian kernel bandwidth}
}
\description{
A simplistic weight function for local regression. This
probably won't be around for long...
}

