\name{local_polynomial_derivatives}
\alias{local_polynomial_derivatives}
\title{Fit local polynomial models to estimate derivatives of a time.table}
\usage{
local_polynomial_derivatives(tt, k, timesteps = NULL, times = NULL, ...)
}
\description{
Fit local polynomial models to estimate derivatives of a
time.table
}

