\name{local_polynomial_fits}
\alias{local_polynomial_fits}
\title{Fit local polynomial models to estimate derivatives}
\usage{
local_polynomial_fits(tt, points, k, timesteps = NULL, times = NULL,
  degree = 2, weight.function = constant_weight_function, vars = NULL)
}
\description{
Fit local polynomial models to estimate derivatives
}

