\name{time_table_lars}
\alias{time_table_lars}
\title{Perform lasso regression on a time.table}
\usage{
time_table_lars(x, y = NULL, idxs = NULL, use.auxiliary = FALSE,
  input.cols = NULL, output.cols = NULL, ..., modelfun = modelfun,
  has.no.na = FALSE)
}
\description{
Perform lasso regression on a time.table
}

